Renamed json_editor to jsoneditor
