+++ panel/models/index.ts
@@ -11,7 +11,7 @@ export {HTML} from "./html"
 export {IDOM} from "./idom"
 export {IPyWidget} from "./ipywidget"
 export {JSON} from "./json"
-export {JSONEditor} from "./json_editor"
+export {JSONEditor} from "./jsoneditor"
 export {FileDownload} from "./file_download"
 export {KaTeX} from "./katex"
 export {Location} from "./location"
